literature_invention_search/requirements.txt
requests>=2.28.0
openai>=1.0.0
anthropic>=0.7.0
biopython>=1.80
python-dotenv>=0.20.0 # For loading .env files
